refactor: move walrus context to types
